feat: add relation between Record & User

+++ routes/modules/record.js
@@ -9,8 +9,11 @@ router.get("/new", (req, res) => {
 
 router.post("/", (req, res) => {
   const record = req.body;
+  // console.log(record);
   Category.findOne({ name: record.categoryId }).then((category) => {
     record.categoryId = category._id;
+    record.userId = req.user._id;
+
     Record.create(record)
       .then((record) => {
         category.record.push(record._id);
@@ -23,33 +26,25 @@ router.post("/", (req, res) => {
 
 router.get("/:id/edit", (req, res) => {
   const _id = req.params.id;
+  const userId = req.user._id;
   //使用populate()
-  Record.findById(_id)
+  Record.findOne({ _id, userId })
     .populate("categoryId")
     .lean()
     .then((record) => {
       res.render("edit", { record });
+      console.log(record);
     })
     .catch((error) => console.log(error));
-
-  //不使用populate()
-  // Record.findById(_id)
-  //   .lean()
-  //   .then((record) => {
-  //     Category.findById(record.categoryId)
-  //       .lean()
-  //       .then((category) => {
-  //         res.render("edit", { record, category });
-  //       })
-  //       .catch((error) => console.log(error));
-  //   });
 });
 
 router.put("/:id", (req, res) => {
   const _id = req.params.id;
+  const userId = req.user._id;
   const update = req.body;
 
-  Record.findOne({ _id })
+  //remove record from old category
+  Record.findOne({ _id, userId })
     .then((record) => {
       Category.findOne(record.categoryId)
         .then((category) => {
@@ -62,11 +57,11 @@ router.put("/:id", (req, res) => {
     })
     .catch((error) => console.log(error));
 
+    //update record
   Category.findOne({ name: update.categoryId })
     .then((category) => {
       update.categoryId = category._id;
-
-      Record.findOneAndUpdate({ _id }, req.body)
+      Record.findOneAndUpdate({ _id, userId }, req.body)
         .then((record) => {
           category.record.push(record._id);
           category.save();
@@ -79,8 +74,10 @@ router.put("/:id", (req, res) => {
 
 router.delete("/:id", (req, res) => {
   const _id = req.params.id;
+  const userId = req.user._id;
 
-  Record.findOne({ _id })
+  //remove record from old category
+  Record.findOne({ _id, userId })
     .then((record) => {
       Category.findOne(record.categoryId)
         .then((category) => {
@@ -93,7 +90,7 @@ router.delete("/:id", (req, res) => {
     })
     .catch((error) => console.log(error));
 
-  Record.findOneAndRemove({ _id })
+  Record.findOneAndRemove({ _id, userId })
     .then(() => res.redirect("/"))
     .catch((error) => console.log(error));
 });
